requirements.txt
aiohttp
cchardet
pyzmq
msgpack-python
click
